GGW-309 added target value as meassure

+++ amp/WEB-INF/src/org/digijava/kernel/ampapi/endpoints/indicator/AmpDashboard/DashboardCoreIndicatorValue.java
@@ -5,15 +5,15 @@ import java.math.BigDecimal;
 public class DashboardCoreIndicatorValue {
     private String indicator;
     private Long indicator_id;
-    private Double targetValue;
+    private BigDecimal targetValue;
     private BigDecimal actualValue;
     private DashboardCoreIndicatorType coreIndicatorType;
 
-    public Double getTargetValue() {
+    public BigDecimal getTargetValue() {
         return targetValue;
     }
 
-    public void setTargetValue(Double targetValue) {
+    public void setTargetValue(BigDecimal targetValue) {
         this.targetValue = targetValue;
     }
